a function that creates a file

+++ 0x15-file_io/1-create_file.c
@@ -14,7 +14,7 @@
 int create_file(const char *filename, char *text_content)
 {
 int descriptor;
-ssize_t num_written;
+int num_written;
 
 if (filename == NULL)
 {
